Use immutable type when performing CO-RE relocations
The use of non immutable types when applying CO-RE relocations causes
quite the memory churn, since a deep copy of the type needs to be
performed. The types from the spec should never be copied though, which
means that all those type copies are not needed.

To fix this, in the `CORERelocate` function, type are now loaded via the
immutable map. This will get rid of the type copies.

This greatly reduces the number of allocations/op when running the
`BenchmarkNewCollectionManyProgs` benchmark. Here are the results:

Previous:
BenchmarkNewCollectionManyProgs-4 19 59275685 ns/op 75121070 B/op 314863 allocs/op

With optimization:
BenchmarkNewCollectionManyProgs-4 754 1627725 ns/op 256659 B/op 4112 allocs/op